requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
llama-cpp-python
huggingface_hub
python-multipart
supabase
